# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Regulate AI - Reporting Bridge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all report catalogues
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published]
        - name: domain
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Create a new report catalogue (draft)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a report catalogue by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Update a report catalogue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdate"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a draft report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /reports/{id}/publish:
    post:
      operationId: publishReport
      tags: [reports]
      summary: Publish a report catalogue to targeted companies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}/compliance:
    get:
      operationId: getReportCompliance
      tags: [reports]
      summary: Get compliance breakdown for a published report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Compliance breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceBreakdown"

  # ── Report Schema Columns ──────────────────────────────────────────────────
...
components:
  schemas:
...
    DomainStat:
      type: object
      properties:
        domain:
          type: string
        complianceRate:
          type: number
        submitted:
          type: integer
        pending:
          type: integer
        overdue:
          type: integer
      required:
        - domain
        - complianceRate
        - submitted
        - pending
        - overdue
...
    ComplianceBreakdown:
      type: object
      properties:
        reportId:
          type: integer
        reportName:
          type: string
        totalNotified:
          type: integer
        pending:
          type: integer
        accepted:
          type: integer
        submitted:
          type: integer
        overdue:
          type: integer
        byDomain:
          type: array
          items:
            $ref: "#/components/schemas/DomainStat"
      required:
        - reportId
        - reportName
        - totalNotified
        - pending
        - accepted
        - submitted
        - overdue
        - byDomain

    # ── Report ─────────────────────────────────────────────────────────────────
    Report:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        frequency:
          type: string
          enum: [ad-hoc, daily, weekly, monthly, quarterly, annually]
        targetDomains:
          type: array
          items:
            type: string
        issueDate:
          type: ["string", "null"]
        implementationWindowDays:
          type: integer
        submissionDeadline:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published]
        rulesBookUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - description
        - frequency
        - targetDomains
        - implementationWindowDays
        - status
        - createdAt
        - updatedAt
    ReportInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        frequency:
          type: string
          enum: [ad-hoc, daily, weekly, monthly, quarterly, annually]
        targetDomains:
          type: array
          items:
            type: string
          minItems: 1
        issueDate:
          type: string
        implementationWindowDays:
          type: integer
          minimum: 1
        submissionDeadline:
          type: string
        rulesBookUrl:
          type: string
      required:
        - name
        - description
        - frequency
        - targetDomains
        - implementationWindowDays
    ReportUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        frequency:
          type: string
          enum: [ad-hoc, daily, weekly, monthly, quarterly, annually]
        targetDomains:
          type: array
          items:
            type: string
        issueDate:
          type: string
        implementationWindowDays:
          type: integer
        submissionDeadline:
          type: string
        rulesBookUrl:
          type: string

    # ── Report Schema Columns ─────────────────────────────────────────────────
...