# lib/api-zod/src/generated/types/complianceBreakdown.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Regulate AI - Reporting Bridge API
 * OpenAPI spec version: 0.1.0
 */
import type { DomainStat } from './domainStat';

export interface ComplianceBreakdown {
  reportId: number;
  reportName: string;
  totalNotified: number;
  pending: number;
  accepted: number;
  submitted: number;
  overdue: number;
  byDomain: DomainStat[];
}