# lib/api-zod/src/generated/types/domainStat.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Regulate AI - Reporting Bridge API
 * OpenAPI spec version: 0.1.0
 */

export interface DomainStat {
  domain: string;
  complianceRate: number;
  submitted: number;
  pending: number;
  overdue: number;
}