# lib/api-zod/src/generated/types/reportUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Regulate AI - Reporting Bridge API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportUpdateFrequency } from './reportUpdateFrequency';

export interface ReportUpdate {
  name?: string;
  description?: string;
  frequency?: ReportUpdateFrequency;
  targetDomains?: string[];
  issueDate?: string;
  implementationWindowDays?: number;
  submissionDeadline?: string;
  rulesBookUrl?: string;
}